feat: update ESLint config, add build script, and refactor TxModal component to use TxUI
