+++ package.json
@@ -20,6 +20,7 @@
 		"dev": "pnpm run dev:staging",
 		"dev:staging": "CLOUDFLARE_ENV=staging vite dev --mode staging --port 5173",
 		"dev:production": "CLOUDFLARE_ENV=production vite dev --mode production --port 5174",
+		"build": "pnpm run build:staging",
 		"build:staging": "CLOUDFLARE_ENV=staging vue-tsc --noEmit && pnpm run lint && vite build --mode staging",
 		"build:production": "CLOUDFLARE_ENV=production vue-tsc --noEmit && pnpm run lint && vite build --mode production",
 		"test": "vitest --run",
