+++ src/components.d.ts
@@ -127,6 +127,7 @@ declare module 'vue' {
     TooltipTrigger: typeof import('./components/ui/tooltip/TooltipTrigger.vue')['default']
     TxModal: typeof import('./components/TxModal.vue')['default']
     TxModalOpPreview: typeof import('./components/TxModalOpPreview.vue')['default']
+    TxUI: typeof import('./components/TxUI.vue')['default']
     ValidateSmartEOA: typeof import('./components/ImportAccountModal/ValidateSmartEOA.vue')['default']
   }
 }
